Flask API/requirements.txt
requests==2.28.1
Pillow==9.3.0
Flask==2.2.2
tensorflow==1.15.0
gunicorn~=20.1.0
protobuf==3.20.*
flask-cors==3.0.10